requirements.txt
aiohttp==3.6.2
asyncpg==0.21.0
Cerberus==1.3.2
loguru==0.5.3
psycopg2-binary==2.8.6
PyYAML==5.3.1
yoyo-migrations==7.2.0